Extract Microphone component
