+++ packages/react-composites/src/composites/CallComposite/components/CallControls.tsx
@@ -4,7 +4,7 @@
 import { concatStyleSets, IButtonStyles, Stack } from '@fluentui/react';
 import { _isInLobbyOrConnecting } from '@internal/calling-component-bindings';
 /* @conditional-compile-remove-from(stable): custom button injection */
-import { ControlBarButton, ControlBarButtonProps } from '@internal/react-components';
+import { ControlBarButton } from '@internal/react-components';
 import {
   BaseCustomStyles,
   CameraButton,
@@ -12,16 +12,12 @@ import {
   ControlBarButtonStyles,
   DevicesButton,
   EndCallButton,
-  MicrophoneButton,
   ParticipantMenuItemsCallback,
   ParticipantsButton,
   ScreenShareButton
 } from '@internal/react-components';
 import React, { useMemo } from 'react';
-import { useLocale } from '../../localization';
 import { usePropsFor } from '../hooks/usePropsFor';
-import { useSelector } from '../hooks/useSelector';
-import { getCallStatus, getLocalMicrophoneEnabled } from '../selectors/baseSelectors';
 import {
   controlButtonBaseStyle,
   devicesButtonWithIncreasedTouchTargets,
@@ -29,6 +25,12 @@ import {
   groupCallLeaveButtonStyle,
   participantButtonWithIncreasedTouchTargets
 } from '../styles/CallControls.styles';
+import {
+  CallControlOptions,
+  CustomCallControlButtonPlacement,
+  CustomCallControlButtonProps
+} from '../types/CallControlOptions';
+import { Microphone } from './buttons/Microphone';
 
 /**
  * @private
@@ -62,169 +64,12 @@ export type CallControlsProps = {
 };
 
 /**
- * Control bar display type for {@link CallComposite}.
- *
- * @public
- */
-export type CallControlDisplayType = 'default' | 'compact';
-
-/**
- * Customization options for the control bar in calling experience.
- *
- * @public
- */
-export type CallControlOptions = {
-  /**
-   * Options to change how the call controls are displayed.
-   * `'compact'` display type will decreases the size of buttons and hide the labels.
-   *
-   * @remarks
-   * If the composite `formFactor` is set to `'mobile'`, the control bar will always use compact view.
-   *
-   * @defaultValue 'default'
-   */
-  displayType?: CallControlDisplayType;
-  /**
-   * Show or Hide Camera Button during a call
-   * @defaultValue true
-   */
-  cameraButton?: boolean;
-  /**
-   * Show or Hide EndCall button during a call.
-   * @defaultValue true
-   */
-  endCallButton?: boolean;
-  /**
-   * Show or Hide Microphone button during a call.
-   * @defaultValue true
-   */
-  microphoneButton?: boolean;
-  /**
-   * Show or Hide Devices button during a call.
-   * @defaultValue true
-   */
-  devicesButton?: boolean;
-  /**
-   * Show, Hide or Disable participants button during a call.
-   * @defaultValue true
-   */
-  participantsButton?: boolean | { disabled: boolean };
-  /**
-   * Show, Hide or Disable the screen share button during a call.
-   * @defaultValue true
-   */
-  screenShareButton?: boolean | { disabled: boolean };
-  /* @conditional-compile-remove-from(stable): custom button injection */
-  /**
-   * Inject custom buttons in the call controls.
-   *
-   * @beta
-   */
-  onFetchCustomButtonProps?: CustomCallControlButtonCallback[];
-};
-
-/* @conditional-compile-remove-from(stable): custom button injection */
-/**
- * Placement for a custom button injected in the {@link CallControls}.
- *
- * 'first': Place the button on the left end (right end in rtl mode).
- * 'afterCameraButton': Place the button on the right (left in rtl mode) of the camera button.
- * ... and so on.
- *
- * It is an error to place the button in reference to another button that has
- * been hidden via a {@link CallControlOptions} field.
- *
- * Multiple buttons placed in the same position are appended in order.
- * E.g., if two buttons are placed 'first', they'll both appear on the left end (right end in rtl mode)
- * in the order provided.
- *
- * @beta
- */
-export type CustomCallControlButtonPlacement =
-  | 'first'
-  | 'last'
-  | 'afterCameraButton'
-  | 'afterEndCallButton'
-  | 'afterMicrophoneButton'
-  | 'afterOptionsButton'
-  | 'afterParticipantsButton'
-  | 'afterScreenShareButton';
-
-/* @conditional-compile-remove-from(stable): custom button injection */
-/**
- * A callback that returns the props to render a custom {@link ControlBarButton}.
- *
- * The response indicates where the custom button should be placed.
- *
- * Performance tip: This callback is only called when either the callback or its arguments change.
- *
- * @beta
- */
-export type CustomCallControlButtonCallback = (
-  args: CustomCallControlButtonCallbackArgs
-) => CustomCallControlButtonProps;
-
-/* @conditional-compile-remove-from(stable): custom button injection */
-/**
- * Arguments for {@link CustomCallControlButtonCallback}.
- *
- * @beta
- */
-export interface CustomCallControlButtonCallbackArgs {
-  /**
-   * Buttons should reduce the size to fit a smaller viewport when `displayType` is `'compact'`.
-   *
-   * @defaultValue `'default'`
-   */
-  displayType?: CallControlDisplayType;
-}
-
-/* @conditional-compile-remove-from(stable): custom button injection */
-/**
- * Response from {@link CustomCallControlButtonCallback}.
- *
- * Includes the props necessary to render a {@link  ControlBarButton} and indication of where to place the button.
- *
- * @beta
- */
-export interface CustomCallControlButtonProps extends ControlBarButtonProps {
-  /**
-   * Where to place the custom button relative to other buttons.
-   */
-  placement: CustomCallControlButtonPlacement;
-}
-
-/**
  * @private
  */
 export const CallControls = (props: CallControlsProps): JSX.Element => {
   const { callInvitationURL, onFetchParticipantMenuItems } = props;
   const options = typeof props.options === 'boolean' ? {} : props.options;
   const compactMode = options?.displayType === 'compact';
-
-  const callStatus = useSelector(getCallStatus);
-  const isLocalMicrophoneEnabled = useSelector(getLocalMicrophoneEnabled);
-  const strings = useLocale().strings.call;
-
-  /**
-   * When call is in Lobby, microphone button should be disabled.
-   * This is due to to headless limitation where a call can not be muted/unmuted in lobby.
-   */
-  const microphoneButtonProps = usePropsFor(MicrophoneButton);
-  if (_isInLobbyOrConnecting(callStatus)) {
-    microphoneButtonProps.disabled = true;
-    // Lobby page should show the microphone status that was set on the local preview/configuration
-    // page until the user successfully joins the call.
-    microphoneButtonProps.checked = isLocalMicrophoneEnabled;
-  }
-  const microphoneButtonStrings = _isInLobbyOrConnecting(callStatus)
-    ? {
-        strings: {
-          tooltipOffContent: strings.microphoneToggleInLobbyNotAllowed,
-          tooltipOnContent: strings.microphoneToggleInLobbyNotAllowed
-        }
-      }
-    : {};
 
   const cameraButtonProps = usePropsFor(CameraButton);
   const screenShareButtonProps = usePropsFor(ScreenShareButton);
@@ -275,16 +120,10 @@ export const CallControls = (props: CallControlsProps): JSX.Element => {
   }
 
   const microphoneButton = options?.microphoneButton !== false && (
-    // tab focus on MicrophoneButton on page load
-    <MicrophoneButton
-      autoFocus
-      data-ui-id="call-composite-microphone-button"
-      {...microphoneButtonProps}
-      showLabel={!compactMode}
+    <Microphone
+      displayType={options?.displayType}
       styles={commonButtonStyles}
-      {...microphoneButtonStrings}
-      /* @conditional-compile-remove-from(stable) meeting-composite control-bar-split-buttons */
-      enableDeviceSelectionMenu={props.splitButtonsForDeviceSelection}
+      splitButtonsForDeviceSelection={props.splitButtonsForDeviceSelection}
     />
   );
