+++ packages/react-composites/src/composites/CallComposite/components/buttons/Microphone.tsx
@@ -1,0 +1,58 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { _isInLobbyOrConnecting } from '@internal/calling-component-bindings';
+import { ControlBarButtonStyles, MicrophoneButton } from '@internal/react-components';
+import React from 'react';
+import { useLocale } from '../../../localization';
+import { usePropsFor } from '../../hooks/usePropsFor';
+import { useSelector } from '../../hooks/useSelector';
+import { getCallStatus, getLocalMicrophoneEnabled } from '../../selectors/baseSelectors';
+import { CallControlDisplayType } from '../../types/CallControlOptions';
+/**
+ * @private
+ */
+export const Microphone = (props: {
+  displayType?: CallControlDisplayType;
+  styles?: ControlBarButtonStyles;
+  splitButtonsForDeviceSelection?: boolean;
+}): JSX.Element => {
+  const { displayType, styles, splitButtonsForDeviceSelection } = props;
+  const microphoneButtonProps = usePropsFor(MicrophoneButton);
+  const callStatus = useSelector(getCallStatus);
+  const isLocalMicrophoneEnabled = useSelector(getLocalMicrophoneEnabled);
+  const strings = useLocale().strings.call;
+
+  /**
+   * When call is in Lobby, microphone button should be disabled.
+   * This is due to to headless limitation where a call can not be muted/unmuted in lobby.
+   */
+  if (_isInLobbyOrConnecting(callStatus)) {
+    microphoneButtonProps.disabled = true;
+    // Lobby page should show the microphone status that was set on the local preview/configuration
+    // page until the user successfully joins the call.
+    microphoneButtonProps.checked = isLocalMicrophoneEnabled;
+  }
+  const microphoneButtonStrings = _isInLobbyOrConnecting(callStatus)
+    ? {
+        strings: {
+          tooltipOffContent: strings.microphoneToggleInLobbyNotAllowed,
+          tooltipOnContent: strings.microphoneToggleInLobbyNotAllowed
+        }
+      }
+    : {};
+
+  // tab focus on MicrophoneButton on page load
+  return (
+    <MicrophoneButton
+      autoFocus
+      data-ui-id="call-composite-microphone-button"
+      {...microphoneButtonProps}
+      showLabel={displayType !== 'compact'}
+      styles={styles}
+      {...microphoneButtonStrings}
+      /* @conditional-compile-remove-from(stable) meeting-composite control-bar-split-buttons */
+      enableDeviceSelectionMenu={splitButtonsForDeviceSelection}
+    />
+  );
+};
